pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openclaw-aws-agent"
version = "0.1.0"
description = "AI agent inspired by OpenClaw, rebuilt on AWS managed services"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-builder>=0.1.0",
    "boto3>=1.35.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "hypothesis>=6.0.0",
    "moto[s3,ssm,scheduler]>=5.0.0",
    "pytest-cov>=4.0.0",
]
infra = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["agent*"]
